# balance_api/openapi/spec.yaml
openapi: 3.0.0
info:
  title: Balance Application API
  description: |
    This API allows writing down marks on a Tic Tac Toe board
    and requesting the state of the board or of individual squares.
  version: 1.0.0

components:
  securitySchemes:
    oauth2:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: balance_api.api.security.decode_token

  parameters:
    AccountId:
      in: path
      name: account_id
      description: Unique Account ID
      required: true
      schema:
        type: integer
        format: int32

    TransactionId:
      in: path
      name: transaction_id
      description: Unique Transaction ID
      required: true
      schema:
        type: integer
        format: int32

    TagId:
      in: path
      name: tag_id
      description: Unique Tag ID
      required: true
      schema:
        type: integer
        format: int32

    AccountIdQuery:
      in: query
      name: account_id
      description: User account id
      required: false
      schema:
        type: integer
        format: int32

    PeriodType:
      in: query
      name: period_type
      description: Period Type
      required: false
      schema:
        type: string
        x-extensible-enum:
          - custom
          - last_month

    PeriodOffset:
      in: query
      name: period_offset
      description: Period Offset
      required: false
      schema:
        type: integer
        format: int32

    StartDate:
      in: query
      name: start_date
      description: Start Date
      required: false
      schema:
        type: string
        format: date-time

    EndDate:
      in: query
      name: end_date
      description: End Date
      required: false
      schema:
        type: string
        format: date-time

    Keywords:
      in: query
      name: keywords
      description: Keywords
      required: false
      schema:
        type: string

  schemas:
    Login:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          example: 'john@doe.com'
          description: User email
        password:
          type: string
          example: 'xyz'
          description: The user password

    User:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          format: int32
          example: 1
          description: User id
        name:
          type: string
          nullable: true
          example: 'John Doe'
          description: The user name

    Account:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 5000
          description: account id
        alias:
          type: string
          nullable: true
          example: 'Santander'
          description: The account alias
        user_id:
          type: integer
          format: int32
          example: 1
          description: The user id
        type:
          type: string
          nullable: true
          x-extensible-enum:
            - CHECKING
            - SAVINGS
            - INVESTMENTS
            - OTHERS
          example: 'checking'
          description: The account type

    Accounts:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/Account'

    Tag:
      type: object
      properties:
        id:
          type: integer
          example: 1
          description: account id
        value:
          type: string
          nullable: true
          example: 'Miscellaneous'
          description: The account tag value

    Tags:
      type: object
      properties:
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'

    Transaction:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 123
          description: transaction id
        date:
          type: string
          nullable: true
          format: date-time
          example: '2019-06-26'
          description: The timestamp when the transaction was realized
        transaction_type:
          type: string
          nullable: true
          x-extensible-enum:
            - EXPENSE
            - INCOME
            - TRANSFER
            - REFUND
            - INVESTMENT
            - IOU
          example: 'EXPENSE'
          description: The type of transaction
        amount:
          type: number
          format: float
          example: 3.15
          description: The amount of the transaction
        account:
          $ref: '#/components/schemas/Account'
        description:
          type: string
          example: "Paypal"
          description: Transaction description
        tag_id:
          type: integer
          format: int32
          description: Tag id description

    Transactions:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        balance:
          type: number
          format: float
          example: 3.15
          description: Transaction sum

    Asset:
      type: object
      properties:
        isin:
          type: string
          example: "DE0005933931"
          description: Asset ISIN
        description:
          type: string
          nullable: true
          example: 'ISHARES CORE DAX UCITS ETF (DE)'
          description: Asset description
        price:
          type: number
          format: float
          example: 3.15
          description: Price this assets is being traded.

    Assets:
      type: object
      properties:
        assets:
          type: array
          items:
            $ref: '#/components/schemas/Asset'

paths:

  /oauth2/token:
    post:
      summary: User sign-in
      operationId: balance_api.api.security.oauth.token
      description: User sign-in
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /users:
    post:
      summary: Creates an user
      operationId: balance_api.api.users.create_user
      description: Creates an user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        201:
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    get:
      summary: Retrieves an user
      security:
        - oauth2:
            - uid
      operationId: balance_api.api.users.me
      description: Retrieves an user
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'


  /accounts:
    post:
      summary: Creates an user account
      security:
        - oauth2:
            - uid
      operationId: balance_api.api.accounts.create_account
      description: Creates an user account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
      responses:
        201:
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'

    get:
      summary: Lists all user accounts
      security:
        - oauth2:
            - uid
      operationId: balance_api.api.accounts.list_accounts
      description: Lists all user accounts
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accounts'

  /accounts/{account_id}:
    get:
      summary: Retrieves an user account
      security:
        - oauth2:
            - uid
      operationId: balance_api.api.accounts.find_account
      description: Retrieves an user account
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'


  /tags:
    get:
      summary: Retrieves the list of tags for a user account
      security:
        - oauth2:
            - uid
      operationId: balance_api.api.tags.list_tags
      description: Retrieves the list of tags for a user account
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tags'

  /tags/{tag_id}:
    get:
      summary: Retrieves a tag for a given id
      security:
        - oauth2:
            - uid
      operationId: balance_api.api.tags.find_tag
      description: Retrieves a tag for a given id
      parameters:
        - $ref: '#/components/parameters/TagId'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tags'

  /transactions:
    post:
      summary: Creates a transaction
      security:
        - oauth2:
            - uid
      operationId: balance_api.api.transactions.create_transaction
      description: Creates a transaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transaction'
      responses:
        201:
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'

    get:
      summary: Retrieves the list of transactions for a user
      security:
        - oauth2:
            - uid
      operationId: balance_api.api.transactions.list_transactions
      description: Retrieves the list of transactions for a user
      parameters:
        - $ref: '#/components/parameters/AccountIdQuery'
        - $ref: '#/components/parameters/PeriodType'
        - $ref: '#/components/parameters/PeriodOffset'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transactions'

  /transactions/{transaction_id}:
    put:
      summary: Updates a transaction
      security:
        - oauth2:
            - uid
      operationId: balance_api.api.transactions.update_transaction
      description: Updates a transaction
      parameters:
        - $ref: '#/components/parameters/TransactionId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transaction'
      responses:
        201:
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'

    delete:
      summary: Deletes a single transaction
      security:
        - oauth2:
            - uid
      operationId: balance_api.api.transactions.delete_transaction
      description: Deletes a single transaction for a given transaction_id
      parameters:
        - $ref: '#/components/parameters/TransactionId'
      responses:
        204:
          description: NO CONTENT


  /upload:
    post:
      summary: Uploads a file with several transactions
      security:
        - oauth2:
            - uid
      operationId: balance_api.api.transactions.upload_transaction
      description: Uploads a file with several transactions
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                account_id:
                  type: string
                userId:
                  type: integer
                file:
                  type: string
                  format: binary
      responses:
        201:
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'

  /assets:
    get:
      summary: Searches for assets based on keywords
      security:
        - oauth2:
            - uid
      operationId: balance_api.api.assets.search_assets
      description: Searches for assets based on keywords
      parameters:
        - $ref: '#/components/parameters/Keywords'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assets'

  /reports/balance:
    get:
      summary: Return the account(s) total balance
      security:
        - oauth2:
            - uid
      operationId: balance_api.api.reports.get_balance
      description: Return the account(s) total balance
      parameters:
        - $ref: '#/components/parameters/AccountIdQuery'
        - $ref: '#/components/parameters/PeriodType'
        - $ref: '#/components/parameters/PeriodOffset'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assets'